[#7] Feat : TicketGenerator type 제작, 적용
