+++ BankManagerConsoleApp/BankManagerConsoleApp/Bank.swift
@@ -8,18 +8,12 @@
 import Foundation
 
 struct Bank {
-    private(set) var ticketNumber: Int = 0
     private(set) var numberOfBankTeller: Int
     private var openTime: CFAbsoluteTime?
     private var closeTime: CFAbsoluteTime?
     
     init(numberOfBankTeller: Int) {
         self.numberOfBankTeller = numberOfBankTeller
-    }
-    
-    mutating func getNewTicketNumber() -> Int {
-        ticketNumber += 1
-        return ticketNumber
     }
     
     mutating func openBank() {
@@ -33,12 +27,8 @@ struct Bank {
               let close = closeTime else { throw BankManagerError.failToCaclulateSpentTime }
         
         let spentTime = formatTimeDuration(time: close - open)
-        print("업무가 마감되었습니다. 오늘 업무를 처리한 고객은 총 \(ticketNumber)명이며, 총 업무시간은 \(spentTime)초입니다.")
-        setTicketNumberToZero()
-    }
-    
-    mutating private func setTicketNumberToZero() {
-        ticketNumber = 0
+        print("업무가 마감되었습니다. 오늘 업무를 처리한 고객은 총 \(TicketGenerator.ticketNumber)명이며, 총 업무시간은 \(spentTime)초입니다.")
+        TicketGenerator.resetTicketNumberToZero()
     }
     
     private func formatTimeDuration(time: CFAbsoluteTime) -> String {
