+++ BankManagerConsoleApp/BankManagerConsoleApp/RandomGenerator.swift
@@ -8,12 +8,12 @@
 import Foundation
 
 struct RandomGenerator {
-    func generateRandomCustomer(bank: inout Bank) -> [Customer] {
+    func generateRandomCustomer() -> [Customer] {
         let totalCustomer = createRandomNumberInRange(10, to: 30)
         var customers = [Customer]()
         
         for _ in 1...totalCustomer {
-            let ticketNumber = bank.getNewTicketNumber()
+            let ticketNumber = TicketGenerator.getNewTicketNumber()
             let priority = Customer.Priority.random
             let task = Customer.Task.random
             let randomCustomer  = Customer(ticketNumber: ticketNumber, priority: priority, task: task)
