+++ BankManagerConsoleApp/ConsoleAppTest/TicketGeneratorTest.swift
@@ -1,0 +1,27 @@
+//
+//  TicketGeneratorTest.swift
+//  ConsoleAppTest
+//
+//  Created by 천수현 on 2021/05/03.
+//
+
+import XCTest
+
+class TicketGeneratorTest: XCTestCase {
+
+    override func setUpWithError() throws {
+        
+    }
+
+    override func tearDownWithError() throws {
+        
+    }
+    
+    func test_getNewTicket() {
+        
+    }
+
+    func test_resetTicketNumber() {
+        
+    }
+}
